Fixed non deleting quest item used in action. Patch provided by serPobre. Change grid error assert fail message to more user friendly view.
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@2221 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/ItemHandler.cpp
@@ -201,7 +201,7 @@ void WorldSession::HandleItemQuerySingleOpcode( WorldPacket & recv_data )
         {
             data << pProto->Spells[s].SpellId;
             data << pProto->Spells[s].SpellTrigger;
-            data << pProto->Spells[s].SpellCharges;
+            data << uint32(pProto->Spells[s].SpellCharges);
             data << pProto->Spells[s].SpellCooldown;
             data << pProto->Spells[s].SpellCategory;
             data << pProto->Spells[s].SpellCategoryCooldown;
